Removed the AwaitKit class to avoid conflicts with the AwaitKit namespace. Moved the methods from the DispatchQueue extensions to the main file because Swift wrongly detects ambiguous use of them. New pod version.
