+++ Example/ViewController.swift
@@ -8,6 +8,7 @@
 
 import UIKit
 import PromiseKit
+import AwaitKit
 
 struct User {
   var name: String
@@ -17,13 +18,13 @@ class ViewController: UIViewController {
   override func viewDidLoad() {
     super.viewDidLoad()
 
-    async {
+    AwaitKit.async {
       throw NSError(domain: "com.yannickloriot.error", code: 3, userInfo: nil)
     }
 
-    let user = try! await(signIn(username: "Foo", password: "Bar"))
-    try! await(sendWelcomeMailToUser(user))
-    try! await(redirectToThankYouScreen())
+    let user = try! AwaitKit.await(signIn(username: "Foo", password: "Bar"))
+    try! AwaitKit.await(sendWelcomeMailToUser(user))
+    try! AwaitKit.await(redirectToThankYouScreen())
     
     print("All done with \(user)!")
   }
