+++ Sources/AwaitKit.swift
@@ -27,11 +27,10 @@
 import Foundation
 import PromiseKit
 
-/// Convenience class to make the background job.
-final class AwaitKit {
-  static let asyncQueue = DispatchQueue(label: "com.yannickloriot.asyncqueue", attributes: .concurrent)
-  static let awaitQueue = DispatchQueue(label: "com.yannickloriot.awaitqueue", attributes: .concurrent)
-}
+let asyncQueue = DispatchQueue(label: "com.yannickloriot.asyncqueue", attributes: .concurrent)
+let awaitQueue = DispatchQueue(label: "com.yannickloriot.awaitqueue", attributes: .concurrent)
+
+// MARK: - Async -
 
 /**
  Yields the execution to the given closure and returns a new promise.
@@ -40,7 +39,7 @@ final class AwaitKit {
  - returns: A new promise that is resolved when the provided closure returned.
  */
 public func async<T>(_ body: @escaping () throws -> T) -> Promise<T> {
-  return AwaitKit.asyncQueue.promise(execute: body)
+    return asyncQueue.promise(execute: body)
 }
 
 /**
@@ -49,8 +48,31 @@ public func async<T>(_ body: @escaping () throws -> T) -> Promise<T> {
  - parameter body: The closure that is executed on a concurrent queue.
  */
 public func async(_ body: @escaping () throws -> Void) {
-  AwaitKit.asyncQueue.async(body)
+    async(body, queue: asyncQueue)
 }
+
+/**
+ Yields the execution to the given closure and returns a new promise.
+ 
+ - Parameter body: The closure that is executed on the given queue.
+ - Returns: A new promise that is resolved when the provided closure returned.
+ */
+public func async<T>(_ body: @escaping () throws -> T, queue: DispatchQueue) -> Promise<T> {
+    return queue.promise(execute: body)
+}
+
+/**
+ Yields the execution to the given closure which returns nothing.
+ 
+ - Parameter body: The closure that is executed on the given queue.
+ */
+public func async(_ body: @escaping () throws -> Void, queue: DispatchQueue) {
+    let promise: Promise<Void> = async(body)
+    
+    promise.catch { _ in }
+}
+
+// MARK: - Await -
 
 /**
  Awaits that the given closure finished and returns its value or throws an error if the closure failed.
@@ -61,7 +83,7 @@ public func async(_ body: @escaping () throws -> Void) {
  */
 @discardableResult
 public func await<T>(_ body: @escaping () throws -> T) throws -> T {
-  return try AwaitKit.awaitQueue.await(body)
+    return try await(body, queue: awaitQueue)
 }
 
 /**
@@ -73,5 +95,62 @@ public func await<T>(_ body: @escaping () throws -> T) throws -> T {
  */
 @discardableResult
 public func await<T>(_ promise: Promise<T>) throws -> T {
-  return try AwaitKit.awaitQueue.await(promise)
+    return try await(promise, queue: awaitQueue)
+}
+
+/**
+ Awaits that the given closure finished on the receiver and returns its value or throws an error if the closure failed.
+ 
+ - parameter body: The closure that is executed on the receiver.
+ - throws: The error sent by the closure.
+ - returns: The value of the closure when it is done.
+ - seeAlso: await(promise:)
+ */
+@discardableResult
+public func await<T>(_ body: @escaping () throws -> T, queue: DispatchQueue) throws -> T {
+    let promise = queue.promise(execute: body)
+    
+    return try await(promise, queue: queue)
+}
+
+/**
+ Awaits that the given promise resolved on the receiver and returns its value or throws an error if the promise failed.
+ 
+ - parameter promise: The promise to resolve.
+ - throws: The error produced when the promise is rejected or when the queues are the same.
+ - returns: The value of the promise when it is resolved.
+ */
+@discardableResult
+public func await<T>(_ promise: Promise<T>, queue: DispatchQueue) throws -> T {
+    guard queue.label != DispatchQueue.main.label else {
+        throw NSError(domain: "com.yannickloriot.awaitkit", code: 0, userInfo: [
+            NSLocalizedDescriptionKey: "Operation was aborted.",
+            NSLocalizedFailureReasonErrorKey: "The current and target queues are the same."
+            ])
+    }
+    
+    var result: T?
+    var error: Error?
+    
+    let semaphore = DispatchSemaphore(value: 0)
+    
+    promise
+        .then(on: queue) { value -> Void in
+            result = value
+            
+            semaphore.signal()
+        }
+        .catch(on: queue) { err in
+            error = err
+            
+            semaphore.signal()
+    }
+    
+    _ = semaphore.wait(timeout: DispatchTime(uptimeNanoseconds: UINT64_MAX))
+    
+    guard let unwrappedResult = result else {
+        throw error!
+    }
+    
+    return unwrappedResult
 }
